fluxion: add service to provide scheduler
This build is a bit tricky, because we want to share
packages for jgf/jobspecs along with protos but do not
want to require flux-sched libs locally to build charts,
etc. The trick is to store them alongside the fluxion build
with underscore directories (so they are ignored on the local
machine) and then move those directories to be found and used
in the Dockerfile. It is imperfect, but an OK solution for now.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>