scheduler and queue: start of work
We need a very simple custom scheduler plugin to receive
and deploy (bind) node assignments, and we will do that
with the fluxqueue-scheduler I am adding here. The queue
will live alongside the controller and interact with
the fluxion service, and I have added the skeleton for that
(which needs a lot of work, but it is building so this
is a great start!

Signed-off-by: vsoch <vsoch@users.noreply.github.com>