feat: support for reservations
Since fluxion is single threaded, we need to ensure the
maxWorkers is 1, and a single job to cancel reservations
is issued at the end of each batch group of jobs (that
are issuing calls to AskFlux to schedule work). For the
final job to clear reservations, we retrieve them from
the reservation table, issue a cancel to fluxion for
each one, and then clear the rows of the table as we go.
This ensures that if there is an error to cancel a job
it will be tried again in the next loop.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>