feat: start of partial cancel
This is the start of work for partial cancel.
Specifically, we needed to refactor the match
and jobspec to ask for resources on the level
of a core for a slot (and not node as previously
done). This means we need to get back the
graph and parse the cores into units based
on the size of the slot (pod).

Signed-off-by: vsoch <vsoch@users.noreply.github.com>